+++ BoardGamesShopMVC.Web/Controllers/PublisherController.cs
@@ -51,9 +51,12 @@ namespace BoardGamesShopMVC.Web.Controllers
             var id = _publisherService.AddPublisher(model);
             return RedirectToAction("Index");
         }
-        public IActionResult Delete(int id)
+        public IActionResult Delete(int id, int boardGamesCount)
         {
+            if(boardGamesCount == 0)
+            {
             _publisherService.DeletePublisher(id);
+            }
             return RedirectToAction("Index");
         }
         [HttpGet]
